% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_image_dummy.R
\name{add_image_dummy}
\alias{add_image_dummy}
\title{A function that adds a dummy variable to the dataset indicating which
   messages contain links to images.}
\usage{
add_image_dummy(dataset, var_full_url = "links_full_url")
}
\arguments{
\item{dataset}{The name of your dataset.}

\item{var_full_url}{A string indicating the name of the 
variable in your dataset that conatins lists with full
urls in the text messages.}
}
\description{
This function takes the URLs in the tweets' text and it
   indicates whether this contains a link to an image (==1) 
   or not (==0).
}
\examples{
dataset_with_image_dummy <- add_links_url(dataset, "links_full_url")

}
\keyword{image}

