% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_link_dummy.R
\name{add_link_dummy}
\alias{add_link_dummy}
\title{A function to add a Link dummy variable.}
\usage{
add_link_dummy(dataset, var_text)
}
\arguments{
\item{dataset}{The name of your dataset.}

\item{var_text}{A string indicating the name of the 
variable in your dataset that conatins text messages.}
}
\description{
This function allows you to add a dummy variable to 
   your dataset to indicate with messages contain links to URLs.
   If the message contains a link, link_dummy == 1.
}
\examples{
dataset_with_link_dummy <- add_link_dummy(orig_dataset, "Contents")

}
\keyword{links}

