% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_links_url.R
\name{add_links_url}
\alias{add_links_url}
\title{A function adds 2 variables in the datset with the "shortened"
   and "complete" links in the messages
   
This function takes the text message in tweets and looks for urls
   in them. Then in creates 2 variables "links_short_url" and 
   "links_full_url" to the datset with the links to those urls.}
\usage{
add_links_url(dataset, var_text)
}
\arguments{
\item{dataset}{The name of your dataset.}

\item{var_text}{A string indicating the name of the 
variable in your dataset that conatins text messages.}
}
\description{
A function adds 2 variables in the datset with the "shortened"
   and "complete" links in the messages
   
This function takes the text message in tweets and looks for urls
   in them. Then in creates 2 variables "links_short_url" and 
   "links_full_url" to the datset with the links to those urls.
}
\examples{
 \dontrun{
dataset_with_links_urls <- add_links_url(dataset, "Contents")
}
}
\keyword{urls}

