% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rt_dummy.R
\name{add_rt_dummy}
\alias{add_rt_dummy}
\title{A function to add a RT dummy variable.}
\usage{
add_rt_dummy(dataset, var_text)
}
\arguments{
\item{dataset}{The name of your dataset.}

\item{var_text}{A string indicating the name of the 
variable in your dataset that conatins text messages.}
}
\description{
This function allows you to add a dummy variable to 
   your dataset to indicate with messages are RTs.
   If the message is a retweet, RT == 1.
}
\examples{
dataset_with_rt_dummy <- add_rt_dummy(orig_dataset, "Contents")
}
\keyword{retweet}

