% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tweet_id.R
\name{add_tweet_id}
\alias{add_tweet_id}
\title{A function to add a variable with the tweet id.}
\usage{
add_tweet_id(dataset, var_tweet_url)
}
\arguments{
\item{dataset}{The name of your dataset.}

\item{var_tweet_url}{A string indicating the name of the variable with the 
tweet URL.}
}
\description{
This function allows you to add a variable with the tweet id.
}
\examples{
 \dontrun{
dataset_with_tweet_id <- add_link_dummy(dataset, "URL")
}
}
\keyword{tweet_id}

