% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraire.R
\name{extraire}
\alias{extraire}
\title{A function to get the long urls in tweets}
\usage{
extraire(entree, motif)
}
\arguments{
\item{entree}{The word that indicates where the long_url starts}

\item{motif}{The word that indicates where the long_url ends}
}
\description{
This function takes the output of the getURL() function
   of the package "RCurl" and it returns the long_url of
   given shortened urls. The credit for this function should
   be given to Arthur Charpentier, who wrote this post:
   http://www.r-bloggers.com/r-twitter-and-urls/
}
\examples{
res <- try(extraire(long_urls,"\\r\\nlocation: (.*?)\\r\\nserver"))


}
\keyword{substring}

